ml_services/ocr_service/requirements.txt
paddleocr==3.2.0
opencv-python-headless
grpcio
grpcio-tools
Pillow
pyspellchecker
albumentations